[Android][Top Seller][GM Stats] Wawasan Pasar https://phab.tokopedia.com/T15756 [BUG] GMSTAT - User can tick more than 1 date range https://phab.tokopedia.com/T15654 [BUG] GMSTAT - "Transaksi Berhasil" is disappear when transaction = 1.000.000 (1jt) https://phab.tokopedia.com/T15778 [BUG] [JSON] GMSTAT - Loading forever if DiffConv not equals 0 https://phab.tokopedia.com/T15830 [BUG] [JSON] GMSTAT - Display "Tidak ada data" when no data available https://phab.tokopedia.com/T15837

+++ sellerapp/src/main/java/com/tokopedia/sellerapp/gmstat/models/GetProductGraph.java
@@ -32,7 +32,7 @@ public class GetProductGraph {
     private Double diffTrans;
     @SerializedName("DiffConv")
     @Expose
-    private Integer diffConv;
+    private Double diffConv;
     @SerializedName("DiffView")
     @Expose
     private Double diffView;
@@ -159,7 +159,7 @@ public class GetProductGraph {
      * @return
      * The diffConv
      */
-    public Integer getDiffConv() {
+    public Double getDiffConv() {
         return diffConv;
     }
 
@@ -168,7 +168,7 @@ public class GetProductGraph {
      * @param diffConv
      * The DiffConv
      */
-    public void setDiffConv(Integer diffConv) {
+    public void setDiffConv(Double diffConv) {
         this.diffConv = diffConv;
     }
